+++ CollabApp/CollabApp.mvc/Validation/StringValidator.cs
@@ -1,32 +1,8 @@
 
+using CollabApp.mvc.Exceptions;
+
 namespace CollabApp.mvc.Validation
 {
-    public class ValidationException : Exception
-    {
-        public ValidationException(string message) : base(message) { }
-    }
-
-    public class EmptyFieldException : ValidationException
-    {
-        public EmptyFieldException() : base("Field cannot be empty.") { }
-    }
-
-    public class MaxLengthExceededException : ValidationException
-    {
-        public MaxLengthExceededException(int maxLength)
-            : base($"String exceeds maximum length of {maxLength} characters.") { }
-    }
-
-    public class InvalidCharException : ValidationException
-    {
-        public InvalidCharException() : base("Found invalid character.") { }
-    }
-
-    public class ProfanityException : ValidationException
-    {
-        public ProfanityException() : base("Profanities are not allowed.") { }
-    }
-
     public enum MaxLengths 
     {
         Description = 256,
